image_edit_annotator/requirements.txt
Flask==2.3.3
Pillow==10.0.1
Werkzeug==2.3.7
ImageHash==4.3.1
numpy==1.24.3